// utils/types.ts
export interface iEnvVariables {
    AUTO_PLAYWRIGHT_DEBUG: boolean;
    BASE_URL: string;
    OPENAI_API_KEY: string;
    SCALAR_DASH409A_PASSWORD: string;
    SCALAR_DASH409A_USERNAME: string;
    SCALAR_PASSWORD: string;
    SCALAR_USERNAME: string;
    TESTRAIL_API_KEY: string;
    TESTRAIL_USERNAME: string;
};